Move session segment modals
