+++ frontend/src/pages/Sessions/SearchSidebar/SegmentModals/DeleteSessionSegmentModal.tsx
@@ -9,7 +9,7 @@ import { message } from 'antd'
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
 
-import { useDeleteSegmentMutation } from '../../../../../graph/generated/hooks'
+import { useDeleteSegmentMutation } from '../../../../graph/generated/hooks'
 
 const NO_SEGMENT = 'none'
 
@@ -21,12 +21,9 @@ interface Props {
 	afterDeleteHandler?: () => void
 }
 
-const DeleteSessionSegmentModal: React.FC<React.PropsWithChildren<Props>> = ({
-	hideModalHandler,
-	showModal,
-	segmentToDelete,
-	afterDeleteHandler,
-}) => {
+export const DeleteSessionSegmentModal: React.FC<
+	React.PropsWithChildren<Props>
+> = ({ hideModalHandler, showModal, segmentToDelete, afterDeleteHandler }) => {
 	const { segment_id, project_id } = useParams<{
 		segment_id: string
 		project_id: string
@@ -85,5 +82,3 @@ const DeleteSessionSegmentModal: React.FC<React.PropsWithChildren<Props>> = ({
 		/>
 	)
 }
-
-export default DeleteSessionSegmentModal
